Revert "Revert "Fix Errors""
This reverts commit d137887d221738c7352a7342b54bd09b5922b470.

+++ src/com/spm/timetablemanagement/form/student_details/enterDp.java
@@ -24,6 +24,9 @@ import javax.swing.table.TableModel;
 public class enterDp extends javax.swing.JPanel {
 
     private Connection connection;
+    private int dpId;
+    PreparedStatement pstDp;
+    ResultSet rs;
     /**
      * Creates new form enterDp
      */
@@ -173,7 +176,15 @@ public class enterDp extends javax.swing.JPanel {
             new String [] {
                 "ID", "Degree Program"
             }
-        ));
+        ) {
+            boolean[] canEdit = new boolean [] {
+                false, false
+            };
+
+            public boolean isCellEditable(int rowIndex, int columnIndex) {
+                return canEdit [columnIndex];
+            }
+        });
         tbl_Dp.setCursor(new java.awt.Cursor(java.awt.Cursor.HAND_CURSOR));
         tbl_Dp.addMouseListener(new java.awt.event.MouseAdapter() {
             public void mouseClicked(java.awt.event.MouseEvent evt) {
@@ -181,6 +192,10 @@ public class enterDp extends javax.swing.JPanel {
             }
         });
         jScrollPane1.setViewportView(tbl_Dp);
+        if (tbl_Dp.getColumnModel().getColumnCount() > 0) {
+            tbl_Dp.getColumnModel().getColumn(0).setResizable(false);
+            tbl_Dp.getColumnModel().getColumn(1).setResizable(false);
+        }
 
         txt_id.setText("ID");
 
@@ -247,6 +262,7 @@ public class enterDp extends javax.swing.JPanel {
         {
             JOptionPane.showMessageDialog(null, e);
         }
+
     }//GEN-LAST:event_btn_addDpActionPerformed
 
     private void tbl_DpMouseClicked(java.awt.event.MouseEvent evt) {//GEN-FIRST:event_tbl_DpMouseClicked
@@ -287,7 +303,9 @@ public class enterDp extends javax.swing.JPanel {
 
     private void btn_deleteDpActionPerformed(java.awt.event.ActionEvent evt) {//GEN-FIRST:event_btn_deleteDpActionPerformed
         // TODO add your handling code here:
+        
         String id = txt_id.getText();
+        int whileWorking = 0;
 
         try
         {
@@ -299,12 +317,47 @@ public class enterDp extends javax.swing.JPanel {
             }
             else{
                 txt_error.setText("");
+             
+            
+            String getDpQuery = "select gId from generated_group_id WHERE dpId = '"+id+"'";
+            pstDp = connection.prepareStatement(getDpQuery);
+            rs = pstDp.executeQuery();
+            while(rs.next())
+            {    
+
+                    int x = JOptionPane.showConfirmDialog(this,"You have this related data,is it ok?", "Confirm", JOptionPane.YES_NO_OPTION);
+                    if (x == 0){
+                        ResultSet rs = smt.executeQuery("select id from generated_group_id where dpId = "+id);
+                        String id_G="";
+                        while(rs.next()){
+                            id_G = rs.getString(1);
+                            System.out.println(id_G);
+                        }
+                        smt.execute("DELETE FROM generated_sub_group_id WHERE gId = '"+id_G+"'");
+                        smt.execute("DELETE FROM generated_group_id WHERE dpId = "+id); 
+                        smt.execute("DELETE FROM degree_program WHERE id = "+id);
+                        smt.execute("DELETE FROM all_details WHERE dp = '"+txt_dp.getText().toString()+"'");
+                        
+                        JOptionPane.showMessageDialog(this, "Record Deleted!");
+
+                        
+                    }
+                    else{
+                        JOptionPane.showMessageDialog(this, "Delete Canceled!");
+                        
+                    }
+                    whileWorking++;
                 
-            smt.execute("DELETE FROM degree_program WHERE id = "+id); 
+            }
+            if(whileWorking == 0){
+                smt.execute("DELETE FROM degree_program WHERE id = "+id);
+                JOptionPane.showMessageDialog(this, "Record Deleted!");
+
+            }
             model.setRowCount(0);
             showDPList();
             txt_dp.setText("");
-            JOptionPane.showMessageDialog(this, "Record Deleted!");
+
             }
         }
         catch(Exception e)
